Changed Email from a package to a library and moved config files from app/config to core/config.
