+++ fuel/core/config/mimes.php
@@ -1,5 +1,8 @@
 <?php
 /*
+ * Mimes.php is from the Codeigniter framework <http://www.codeigniter.com/>
+ */
+/*
 | -------------------------------------------------------------------
 | MIME TYPES
 | -------------------------------------------------------------------
