--- fuel/packages/email/README
@@ -1,3 +1,0 @@
-Fuel Email package v0.1 by DudeAmI (aka Kris)
-
-View the license file for licensing (Licensed under GPL v3.0)
